+++ content-scripts/inject/module.js
@@ -12,7 +12,7 @@ scratchAddons.eventTargets = {
 scratchAddons.session = {};
 const consoleOutput = (logAuthor = "[page]") => {
   const style = {
-    // Remember to change these as well on cs.js
+    // Remember to change these as well on cs-console.js
     leftPrefix: "background:  #ff7b26; color: white; border-radius: 0.5rem 0 0 0.5rem; padding: 0 0.5rem",
     rightPrefix:
       "background: #222; color: white; border-radius: 0 0.5rem 0.5rem 0; padding: 0 0.5rem; font-weight: bold",
