+++ content-scripts/prototype-handler.js
@@ -1,29 +1,31 @@
-function injectPrototype() {
-  const oldBind = Function.prototype.bind;
-  // Use custom event target
-  window.__scratchAddonsTraps = new EventTarget();
-  const onceMap = (__scratchAddonsTraps._onceMap = Object.create(null));
+/* global shouldRun */
 
-  Function.prototype.bind = function (...args) {
-    if (Function.prototype.bind === oldBind) {
-      // Just in case some code stores the bind function once on startup, then always uses it.
-      return oldBind.apply(this, args);
-    } else if (
-      args[0] &&
-      Object.prototype.hasOwnProperty.call(args[0], "editingTarget") &&
-      Object.prototype.hasOwnProperty.call(args[0], "runtime")
-    ) {
-      onceMap.vm = args[0];
-      // After finding the VM, return to previous Function.prototype.bind
-      Function.prototype.bind = oldBind;
-      return oldBind.apply(this, args);
-    } else {
-      return oldBind.apply(this, args);
-    }
-  };
-}
+if (shouldRun && location.pathname.split("/")[1] === "projects") {
+  function injectPrototype() {
+    const oldBind = Function.prototype.bind;
+    // Use custom event target
+    window.__scratchAddonsTraps = new EventTarget();
+    const onceMap = (__scratchAddonsTraps._onceMap = Object.create(null));
+  
+    Function.prototype.bind = function (...args) {
+      if (Function.prototype.bind === oldBind) {
+        // Just in case some code stores the bind function once on startup, then always uses it.
+        return oldBind.apply(this, args);
+      } else if (
+        args[0] &&
+        Object.prototype.hasOwnProperty.call(args[0], "editingTarget") &&
+        Object.prototype.hasOwnProperty.call(args[0], "runtime")
+      ) {
+        onceMap.vm = args[0];
+        // After finding the VM, return to previous Function.prototype.bind
+        Function.prototype.bind = oldBind;
+        return oldBind.apply(this, args);
+      } else {
+        return oldBind.apply(this, args);
+      }
+    };
+  }
 
-if (!(document.documentElement instanceof SVGElement) && location.pathname.split("/")[1] === "projects") {
   const injectPrototypeScript = document.createElement("script");
   injectPrototypeScript.append(document.createTextNode("(" + injectPrototype + ")()"));
   (document.head || document.documentElement).appendChild(injectPrototypeScript);
